Revert "fix: stop using typescript in renderprops API"
This reverts commit d8e8ff01369f3d40cf59e9f8ce7867b3be927b7f.
