--- packages/core/src/components/Spring.jsx
@@ -1,5 +1,0 @@
-import { useSpring } from '../hooks'
-
-export function Spring({ children, ...props }) {
-  return children(useSpring(props))
-}
