--- packages/core/src/components/Transition.jsx
@@ -1,6 +1,0 @@
-import React from 'react'
-import { useTransition } from '../hooks'
-
-export function Transition({ items, children, ...props }) {
-  return <>{useTransition(items, props)(children)}</>
-}
