--- packages/core/src/components/index.js
@@ -1,3 +1,0 @@
-export { Spring } from './Spring'
-export { Trail } from './Trail'
-export { Transition } from './Transition'
