+++ packages/core/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './Spring'
+export * from './Trail'
+export * from './Transition'
